[remark-wiki-link][m]: refactor to typescript
