--- packages/remark-wiki-link/src/index.js
@@ -1,2 +1,0 @@
-export * from "./lib/remark-wiki-link.js";
-export { default } from "./lib/remark-wiki-link.js";
