+++ packages/remark-wiki-link/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./lib/remarkWikiLink";
+export { default } from "./lib/remarkWikiLink";
