+++ packages/remark-wiki-link/src/lib/getPermalinks.ts
